requirements.txt
aiohttp==3.9.5
discord.py==2.5.2
python-dotenv==1.0.1
Flask==3.0.2
pymongo[srv]==4.7.0
dnspython==2.4.2
pyjokes==0.6.0
PyYAML==6.0.1
requests==2.31.0
google-generativeai==0.7.0
google-api-python-client==2.130.0